Move multi signature logic to execution store library
`ExecutionStore` library will contain:
- `Executions` struct: holds the keys, thresholds, and executions
- ERC725 key setters, execute and approve functions

Library funciton modifiers will change to public so that the
compiler will use `DELEGATECALL` instead of `JUMP`. This is to
further decrease gas cost of contract deployment.

Constants for the key purposes and key types will move to a
separate contract which will be used to initialize the registry,
which will in turn use it to initialize the identity contracts.