Implement linking of ethereum addresses to identity
IdentityAccounts is an interface for adding and removing accounts from
an identity. Accounts are blockchain agnostic and is represented in
bytes32.

EthereumIdentityAccounts is a contract that implements IdentityAccounts
and is specific to linking ethereum addresses to an identity contract.

Each identity contract can have multiple ethereum accounts, but
each ethereum account can only belong to one identity contract.

The IdentityRegistry will store an blockchain specific implementation
of IdentityAccounts interface, this is to tie the identity contract
to the various accounts on the other chains, i.e. Stellar.

EthereumIdentityAccounts is a special case in the IdentityRegistry in
that every new identity contract that is created by the
IdentityRegistry will automatically have the identity contract linked
with the account who created the identity contract.